today songah did work everything

+++ enjoytrip-spring/src/main/java/com/ssafy/enjoytrip/qna_comment/model/QnaCommentRequestDto.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 public class QnaCommentRequestDto {
 
     private Long id;           // 댓글 번호 (PK)
-    private Long postId;       // 게시글 번호 (FK)
+    private int postId;       // 게시글 번호 (FK)
     private String content;    // 내용
     private String writer;     // 작성자
